[REFACTOR] analyses 내 task.py 분할 작업
